chore: move tsup config to configs
