+++ packages/interactions/tsup.config.ts
@@ -1,3 +1,3 @@
-import defaultConfig from "../../tsup.config";
+import defaultConfig from "../../configs/tsup.config";
 
 export default defaultConfig;
